@@ -3,10 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "axios": "^0.21.1",
     "create-react-app": "^1.4.3",
-    "prop-types": "^15.6.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "prop-types": "^15.7.2",
+    "react-dom": "^16.14.0",
     "react-scripts": "1.0.17",
     "save": "^2.3.1"
   },
@@ -18,9 +18,10 @@
   "devDependencies": {
     "chai": "^4.1.2",
     "enzyme": "^3.2.0",
-    "enzyme-adapter-react-16": "^1.1.0",
+    "enzyme-adapter-react-16": "^1.15.6",
     "expect": "^21.2.1",
     "jest-junit": "^3.3.0",
+    "react": "^17.0.1",
     "react-test-renderer": "^16.2.0"
   }
 }
